requirements.txt
alembic==1.4.3
cachelib==0.1.1
click==7.1.2
DateTime==4.3
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Authorize==0.2.2
Flask-Dropzone==1.5.4
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
idna==3.1
IPTCInfo3==2.1.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
Pillow==8.1.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.2.0
